+++ src/entity/Player.java
@@ -50,14 +50,14 @@ public class Player extends Entity {
 
         hitArea = new Rectangle();
         boundingBox = new Rectangle();
-        boundingBox.width = 32 * GamePanel.scale;
-        boundingBox.height = 32 * GamePanel.scale;
+        boundingBox.width = 32 * (int)GamePanel.scale;
+        boundingBox.height = 32 * (int)GamePanel.scale;
         hitArea.x = boundingBox.width / 4;
         hitArea.y = boundingBox.height / 3 * 2;
         hitArea.width = boundingBox.width / 2;
         hitArea.height = boundingBox.height / 3;
-        screenX = GamePanel.screenWidth / 2 - boundingBox.width / 2;
-        screenY = GamePanel.screenHeight / 2 - boundingBox.height / 2;
+        screenX = (int) (GamePanel.screenWidth / 2 - boundingBox.width / 2);
+        screenY = (int) (GamePanel.screenHeight / 2 - boundingBox.height / 2);
         boundingBox.x = screenX;
         boundingBox.y = screenY;
         speed = 4;
